Split up `bdd_builder` into pieces
