+++ src/builder/sdd/compression.rs
@@ -191,7 +191,7 @@ impl<'a> CompressionSddBuilder<'a> {
 // check that (a \/ b) /\ a === a
 #[test]
 fn simple_equality() {
-    use crate::builder::bdd_builder::VarLabel;
+    use crate::repr::var_label::VarLabel;
     let builder = CompressionSddBuilder::new(VTree::even_split(
         &[
             VarLabel::new(0),
@@ -213,7 +213,7 @@ fn simple_equality() {
 // check that (a \/ b) | !b === a
 #[test]
 fn sdd_simple_cond() {
-    use crate::builder::bdd_builder::VarLabel;
+    use crate::repr::var_label::VarLabel;
     let builder = CompressionSddBuilder::new(VTree::even_split(
         &[
             VarLabel::new(0),
@@ -240,7 +240,7 @@ fn sdd_simple_cond() {
 
 #[test]
 fn sdd_test_exist() {
-    use crate::builder::bdd_builder::VarLabel;
+    use crate::repr::var_label::VarLabel;
     let builder = CompressionSddBuilder::new(VTree::even_split(
         &[
             VarLabel::new(0),
@@ -269,7 +269,7 @@ fn sdd_test_exist() {
 
 #[test]
 fn sdd_bigand() {
-    use crate::builder::bdd_builder::VarLabel;
+    use crate::repr::var_label::VarLabel;
     let builder = CompressionSddBuilder::new(VTree::right_linear(&[
         VarLabel::new(0),
         VarLabel::new(1),
@@ -306,7 +306,7 @@ fn sdd_bigand() {
 
 #[test]
 fn sdd_ite1() {
-    use crate::builder::bdd_builder::VarLabel;
+    use crate::repr::var_label::VarLabel;
     let builder = CompressionSddBuilder::new(VTree::even_split(
         &[
             VarLabel::new(0),
@@ -336,7 +336,7 @@ fn sdd_ite1() {
 
 #[test]
 fn sdd_demorgan() {
-    use crate::builder::bdd_builder::VarLabel;
+    use crate::repr::var_label::VarLabel;
     let builder = CompressionSddBuilder::new(VTree::even_split(
         &[
             VarLabel::new(0),
@@ -361,7 +361,7 @@ fn sdd_demorgan() {
 
 #[test]
 fn sdd_circuit1() {
-    use crate::builder::bdd_builder::VarLabel;
+    use crate::repr::var_label::VarLabel;
     let builder = CompressionSddBuilder::new(VTree::even_split(
         &[
             VarLabel::new(0),
@@ -394,7 +394,7 @@ fn sdd_circuit1() {
 
 #[test]
 fn sdd_circuit2() {
-    use crate::builder::bdd_builder::VarLabel;
+    use crate::repr::var_label::VarLabel;
     // same as circuit1, but with a different variable order
     let builder = CompressionSddBuilder::new(VTree::even_split(
         &[
@@ -425,7 +425,7 @@ fn sdd_circuit2() {
 
 #[test]
 fn sdd_wmc1() {
-    use crate::builder::bdd_builder::VarLabel;
+    use crate::repr::var_label::VarLabel;
     use crate::util::semirings::realsemiring::RealSemiring;
     // modeling the formula (x<=>fx) && (y<=>fy), with f weight of 0.5
