+++ tests/network_example.rs
@@ -3,7 +3,7 @@
 
 // extern crate rsdd;
 // use crate::repr::var_label::VarLabel;
-// use rsdd::builder::bdd_builder::StandardBddBuilder;
+// use rsdd::builder::bdd_builder::RobddBuilder;
 // use rsdd::builder::cache::all_app::AllTable;
 // use rsdd::repr::robdd::BddPtr;
 // use rsdd::repr::ddnnf::DDNNFPtr;
@@ -17,10 +17,10 @@
 // // For example, network_gen(4) will generate a network with top path edges
 // // st1, t1t2, t2t3, t3t4, t4e.
 // // Output is: BDDPtr to network, associated BDDManager, Vec of edge VarLabels
-// fn network_gen(n: usize) -> (BddPtr, StandardBddBuilder<AllTable<BddPtr>>, Vec<VarLabel>) {
+// fn network_gen(n: usize) -> (BddPtr, RobddBuilder<AllTable<BddPtr>>, Vec<VarLabel>) {
 //     // Initialize BDD
 //     let bdd_size = (2 * n) + 2;
-//     let mut man = StandardBddBuilder::<AllTable<BddPtr>>::new_default_order(bdd_size);
+//     let mut man = RobddBuilder::<AllTable<BddPtr>>::new_default_order(bdd_size);
 
 //     // Initialize variables in the BDD, sort into top or bottom edge
 //     let v: Vec<VarLabel> = (0..bdd_size).map(|x| VarLabel::new(x as u64)).collect();
@@ -51,10 +51,10 @@
 // //         Vec of VarLabels of Decisions, VarLabel of Reward.
 // fn decisions(
 //     n: usize,
-//     mut man: StandardBddBuilder<AllTable<BddPtr>>,
+//     mut man: RobddBuilder<AllTable<BddPtr>>,
 // ) -> (
 //     BddPtr,
-//     StandardBddBuilder<AllTable<BddPtr>>,
+//     RobddBuilder<AllTable<BddPtr>>,
 //     Vec<VarLabel>,
 //     VarLabel,
 // ) {
